#769 Move reader parameters parser to the parser module to support non-Spark readers.
