+++ spark-cobol/src/test/scala/za/co/absa/cobrix/spark/cobol/source/DefaultSourceSpec.scala
@@ -26,7 +26,7 @@ import za.co.absa.cobrix.spark.cobol.source.utils.SourceTestUtils
 class DefaultSourceSpec extends SparkCobolTestBase {
    
   import SourceTestUtils._
-  import za.co.absa.cobrix.spark.cobol.parameters.CobolParametersParser._
+  import za.co.absa.cobrix.cobol.reader.parameters.CobolParametersParser._
   
   private var defaultSource: DefaultSource = _
